webapi/webapi-webrtc-w3c-tests/webrtc/RTCIceCandidate_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu,Yun <yunx.liu@intel.com>

-->

<meta charset="utf-8">
<title>WebRTC Test: RTCIceCandidate basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcicecandidate-type">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>

var initializer = {candidate: "foo", sdpMid: "bar", sdpMLineIndex: 6};
var candidate = new RTCIceCandidate(initializer);

test(function () {
  assert_true("candidate" in candidate, "candidate attribute in RTCIceCandidate");
  assert_equals(Object.prototype.toString.call(candidate.candidate), "[object String]", "candidate attribute of type");
}, "Check if RTCIceCandidate.candidate exists and type of string");

test(function () {
  assert_true("sdpMid" in candidate, "sdpMid attribute in RTCIceCandidate");
  assert_equals(Object.prototype.toString.call(candidate.sdpMid), "[object String]", "sdpMid attribute of type");
}, "Check if RTCIceCandidate.sdpMid exists and type of string");

test(function () {
  assert_true("sdpMLineIndex" in candidate, "sdpMLineIndex attribute in RTCIceCandidate");
  assert_equals(Object.prototype.toString.call(candidate.sdpMLineIndex), "[object Number]", "sdpMLineIndex attribute of type");
}, "Check if RTCIceCandidate.sdpMLineIndex exists and type of number");

</script>
